Update main.rs
Implemented bubble sort and made the sorter use that. With the memoisation addition by delma it seems to perform similarily to the default sort_by (when it's doing insertion sort). This should be a more pleasant experience to the user since every element of the list will be encountered in order in the first n comparisons. There might be a bug in the sort if the list is empty or 1 large but I will fix that later(because it's late) as it's not a critical bug for this usecase.